Add events and refactor/remove various features
Added events that can be assigned callbacks by using the `on` method.

Added the ability to combine character sets by providing an array of values to the 'characterSet' option.

Removed the feature that allows the insertion of garbled text to a given element in favour of passing it via the 'garble' event' callback.

Removed the feature that allows for the listening for an event and calling of `stop` once fired in favour of `stop` being called manually whenever needed.

Removed the feature that allows for the specification of a duration that will call `stop` once elapsed for the same reason as above.

Removed the feature that allows for various transition effects as this was out-of-scope and goes against the design of the module.

Removed the following parameters from TextGarbler:
  - duration
  - stopOn
  - transition

Renamed 'trueValue' to 'baseString'.

Renamed `generateGarbledString` to `garble`.

Renamed 'isRunning' to 'active'.

Moved `generateRandomInteger` outside of TextGarbler's definition.

Moved `generateRandomCharacter` outside of TextGarbler's definition.

Fixed `stop` and `start` to only execute if the text is not currently being garbled/is currently being garbled, respectively.

Refactored the CJK character sets into three seperate sets (Chinese, Japanese, Korean). These sets combine to form the CJK set.

Documentation comments are up-to-date.